move last remaining config-class (which I forgot during the last commit)
